The following files have been updated:
I was working on integrations of  operations connected with   ERC721. Seems we  reached to the point where we could buy, sell ERC721 token. Now it says that we dont have enough  fund to pay for price and gas. We are working on test network setup. We need to rebuild etherlinker plugin because we did changes on setup files. Yesterday we made some progress. Their interaction server is written via Java. We need to do changes in local environment to rebuild server. Also we need to use opportunities of   Kovan

Task Name: Blockchain Integration

Time : 11 Hours.